feat: compose nav menu
