+++ src/assets/icons/IlluminateFantasyLeagueLogo.tsx
@@ -4,8 +4,7 @@ function IlluminateFantasyLeagueLogo() {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
-      width="98"
-      height="34"
+      height="32"
       fill="none"
       viewBox="0 0 467 165"
     >
